XF-368 | accumulate relevant methods and vars in wiki-urls service
