+++ app/models/user.js
@@ -12,7 +12,7 @@ import {getFetchErrorMessage, UserLoadDetailsFetchError, UserLoadInfoFetchError}
 export default EmberObject.extend({
 	defaultAvatarSize: 100,
 	logger: service(),
-	buildUrl: service(),
+	wikiUrls: service(),
 
 	getUserId(accessToken) {
 		if (!accessToken) {
@@ -100,7 +100,7 @@ export default EmberObject.extend({
 	 * @returns {RSVP.Promise}
 	 */
 	loadDetails(host, userId, avatarSize) {
-		const url = this.get('buildUrl').build({
+		const url = this.get('wikiUrls').build({
 			host,
 			path: '/wikia.php',
 			query: {
@@ -142,7 +142,7 @@ export default EmberObject.extend({
 	 * @returns {RSVP.Promise<QueryUserInfoResponse>}
 	 */
 	loadUserInfo(host, accessToken, userId) {
-		const url = this.get('buildUrl').build({
+		const url = this.get('wikiUrls').build({
 			host,
 			path: '/api.php',
 			query: {
@@ -182,7 +182,7 @@ export default EmberObject.extend({
 		const data = {
 			name: userData.name,
 			avatarPath: userData.avatar,
-			profileUrl: this.get('buildUrl').build({
+			profileUrl: this.get('wikiUrls').build({
 				namespace: 'User',
 				relative: true,
 				title: userData.name
