+++ app/models/wikia-nav.js
@@ -2,13 +2,15 @@ import {inject as service} from '@ember/service';
 import {reads, or, bool} from '@ember/object/computed';
 import {A} from '@ember/array';
 import EmberObject, {get, computed} from '@ember/object';
-import {langPathRegexp} from '../utils/language';
 
 export default EmberObject.extend({
 	i18n: service(),
 	logger: service(),
+	wikiUrls: service(),
 	wikiVariables: service(),
+
 	dsGlobalNavigation: null,
+
 	hubsLinks: computed(function () {
 		return this.get('dsGlobalNavigation.fandom_overview.links');
 	}),
@@ -164,7 +166,7 @@ export default EmberObject.extend({
 			this.get('discussionsEnabled') &&
 			[{
 				type: 'nav-menu-external',
-				href: '/d/f',
+				href: `${this.get('wikiUrls.langPath')}/d/f`,
 				name: this.get('i18n').t('app.discussions-label'),
 				trackCategory: 'discussion',
 				trackLabel: 'local-nav'
@@ -176,7 +178,7 @@ export default EmberObject.extend({
 			this.get('currentLocalLinks').map((item, index) => {
 				return {
 					type: item.children ? 'nav-menu-root' : 'nav-menu-item',
-					href: item.href.replace(new RegExp(`^(${langPathRegexp}?/wiki)?/`, 'i'), ''),
+					href: item.href.replace(new RegExp(`^(${this.get('wikiUrls.langPathRegexp')}?/wiki)?/`, 'i'), ''),
 					route: 'wiki-page',
 					name: item.text,
 					index: index + 1,
