+++ tests/unit/components/wikia-search-test.js
@@ -6,9 +6,9 @@ import {module, test} from 'qunit';
 import {setupTest} from 'ember-qunit';
 
 const trackModule = require('mobile-wiki/utils/track');
-let trackStub, component, buildUrlStub;
+let trackStub, component, wikiUrlsBuildStub;
 
-const buildUrlServiceStub = Service.extend({
+const wikiUrlsServiceStub = Service.extend({
 	build() {}
 });
 
@@ -21,13 +21,13 @@ module('Unit | Component | local wikia search', (hooks) => {
 		component.set('wikiVariables', {
 			host: 'wikia.com'
 		});
-		this.owner.register('service:buildUrl', buildUrlServiceStub);
-		const buildUrlService = this.owner.lookup('service:buildUrl');
-		buildUrlStub = sinon.stub(buildUrlService, 'build');
+		this.owner.register('service:wikiUrls', wikiUrlsServiceStub);
+		const wikiUrlsService = this.owner.lookup('service:wikiUrls');
+		wikiUrlsBuildStub = sinon.stub(wikiUrlsService, 'build');
 	});
 
 	hooks.afterEach(() => {
-		buildUrlStub.restore();
+		wikiUrlsBuildStub.restore();
 		trackStub.restore();
 	});
 
@@ -41,7 +41,7 @@ module('Unit | Component | local wikia search', (hooks) => {
 		queries.forEach((query) => {
 			component.getSearchURI(query);
 			assert.ok(
-				buildUrlStub.calledWith({
+				wikiUrlsBuildStub.calledWith({
 					host: 'wikia.com',
 					path: '/wikia.php',
 					query: {
