+++ tests/unit/models/article-comments-test.js
@@ -4,9 +4,9 @@ import {setupTest} from 'ember-qunit';
 import sinon from 'sinon';
 import {run} from '@ember/runloop';
 
-let stub;
+let wikiUrlsBuildStub;
 
-const buildUrlServiceStub = Service.extend({
+const wikiUrlsServiceStub = Service.extend({
 	build() {}
 });
 
@@ -14,13 +14,13 @@ module('Unit | Model | article comments', (hooks) => {
 	setupTest(hooks);
 
 	hooks.beforeEach(function () {
-		this.owner.register('service:buildUrl', buildUrlServiceStub);
-		const buildUrlService = this.owner.lookup('service:buildUrl');
-		stub = sinon.stub(buildUrlService, 'build');
+		this.owner.register('service:wikiUrls', wikiUrlsServiceStub);
+		const wikiUrlsService = this.owner.lookup('service:wikiUrls');
+		wikiUrlsBuildStub = sinon.stub(wikiUrlsService, 'build');
 	});
 
 	hooks.afterEach(() => {
-		stub.restore();
+		wikiUrlsBuildStub.restore();
 	});
 
 	test('url creates valid url to a resource', function (assert) {
@@ -28,7 +28,7 @@ module('Unit | Model | article comments', (hooks) => {
 
 		model.set('host', 'wikia.com');
 		model.url(1, 1);
-		assert.ok(stub.calledWith({
+		assert.ok(wikiUrlsBuildStub.calledWith({
 			host: 'wikia.com',
 			path: '/wikia.php',
 			query: {
